Now loading .obj files with tiny_obj_loader

+++ trace/src/render/Model.h
@@ -19,7 +19,7 @@ namespace trace {
 		~Model();
 
 		void Init(const std::vector<Vertex>& data, const std::vector<uint32_t>& indices);
-		uint32_t GetIndexCount() { return m_indices.size(); }
+		uint32_t GetIndexCount() { return static_cast<uint32_t>(m_indices.size()); }
 		GBuffer* GetIndexBuffer() { return m_indexBuffer; }
 		GBuffer* GetVertexBuffer() { return m_vertexBuffer; }
